added to footer a njump.me home link and a seperate & open-sourced on github link
The njump.me website was missing a "home" link to root. This commit adds it to the footer in the "powered by [njump]()" message, that previously redirected to the github page.

The github link is now added in an "open-sourced on [github]()" message. The new footer is on the format:

"powered by njump & open-sourced on github"

Now linking to both home and the github page.